% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_data.R
\name{create_functional_anova_data}
\alias{create_functional_anova_data}
\title{Create functional data for ANOVA by fitting b-splines to multiple conditions.}
\usage{
create_functional_anova_data(data, pupil, basis = 10, order = 4)
}
\arguments{
\item{data}{a Pupil_data dataframe}

\item{pupil}{column name indicating pupil data to fit}

\item{basis}{number of basis functions to use in fitting}

\item{order}{order of the b-splines (default is cubic splines, order = 4)}
}
\value{
A Pupil_function_data dataframe
}
\description{
This function fits b-splines to pupil data for multiple conditions in preparation for a functional ANOVA.
Unlike the difference-based approach for t-tests, this preserves the original condition structure.
}
\examples{
Sdata <- make_pupillometryr_data(data = pupil_data,
                               subject = ID,
                               trial = Trial,
                               time = Time,
                               condition = Type)
regressed_data <- regress_data(data = Sdata, pupil1 = RPupil, pupil2 = LPupil)
mean_data <- calculate_mean_pupil_size(data = regressed_data, pupil1 = RPupil, pupil2 = LPupil)
base_data <- baseline_data(data = mean_data, pupil = mean_pupil, start = 0, stop = 100)
anova_spline_data <- create_functional_anova_data(data = base_data, pupil = mean_pupil, basis = 10, order = 4)

}
