% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{create_time_windows}
\alias{create_time_windows}
\title{Make PupillometryR dataframe into multiple time windows for easy analysis}
\usage{
create_time_windows(data, pupil, breaks)
}
\arguments{
\item{data}{a PupillometryR dataframe}

\item{pupil}{column name denoting pupil data to be used}

\item{breaks}{a vector or numbers indicating start times for each window}
}
\value{
a Pupil_window_data dataframe
}
\description{
This function creates a single collapsed data frame for easy analysis with an anova or model,
per condition.
By comparison create_window_data allows collapsing all into a single time window.
}
\examples{
window <- create_window_data(data = base_data, pupil = mpupil)
p <- plot(data = window, pupil = mpupil, windows = F, geom = 'boxplot')
p
}
