% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\docType{data}
\name{GeomFlatViolin}
\alias{GeomFlatViolin}
\title{geom_flat_violin_HELPER2}
\description{
Borrowed from
\href{https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R}{Ben Marwick}.
Original author David Robinson.
}
\keyword{datasets}
