% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{gfv_helper1}
\alias{gfv_helper1}
\alias{\%||\%}
\title{geom_flat_violin_HELPER1}
\description{
Borrowed from
https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R.
Original author David Robinson, from https://gist.github.com/dgrtwo/eb7750e74997891d7c20
}
