% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pupil_data.R
\name{plot.Pupil_anova_fda}
\alias{plot.Pupil_anova_fda}
\title{Plot method for Pupil_anova_fda objects}
\usage{
\method{plot}{Pupil_anova_fda}(
  x,
  show_divergence = TRUE,
  use_adjusted = TRUE,
  colour = "black",
  fill = "grey",
  smoothed = FALSE,
  ...
)
}
\arguments{
\item{x}{A Pupil_anova_fda object}

\item{show_divergence}{Logical; whether to show divergence regions}

\item{use_adjusted}{Logical; whether to use FDR/permutation-adjusted divergence}

\item{colour}{Line color for F statistic}

\item{fill}{Fill color for divergence regions}

\item{...}{Additional parameters passed to plotting functions}
}
\value{
A ggplot2 object
}
\description{
Plot method for Pupil_anova_fda objects
}
