% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pupil_data.R
\name{plot.Pupil_test_data}
\alias{plot.Pupil_test_data}
\title{Pre-prepared plots of PupillometryR data}
\usage{
\method{plot}{Pupil_test_data}(x, show_divergence = TRUE,
  colour = "black", fill = "grey", ...)
}
\arguments{
\item{x}{A Pupil_test_data dataframe}

\item{show_divergence}{logical indicating whether divergences are to be highlighted}

\item{colour}{string indicating colour of geom_line, passed to ggplot2}

\item{fill}{string indicating fill hue of divergence highlights, passed to ggplot2}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
The plot functions are designed to run with just data and pupil selections,
with some additional options for fun with plotting. To see these plots,
you must first use one of the run_functional tests.
}
\examples{
ft_data <- run_functional_t_test(data = spline_data,
pupil = mpupil)
p <- plot(ft_data, show_divergence = T, colour = 'red', fill = 'orange')
p
}
