% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pupil_data.R
\name{plot.Pupil_window_data}
\alias{plot.Pupil_window_data}
\title{Pre-prepared plots of PupillometryR data}
\usage{
\method{plot}{Pupil_window_data}(data, pupil, windows = c(FALSE, TRUE),
  geom = c("raincloud", "violin", "boxplot"), ...)
}
\arguments{
\item{data}{A Pupil_window_data dataframe}

\item{pupil}{Column name of pupil data to be plotted}

\item{windows}{Whether you want to include time windows in the plot - logical}

\item{geom}{violin plots or boxplots. The newest version adds raincloud plots using Ben Marwick's flat violin plot.}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
The plot functions are designed to run with just data and pupil selections,
with some additional options for fun with plotting. To see these plots,
you must first use create_window_data.
}
\examples{
window <- create_window_data(data = base_data,pupil = mpupil)
p <-plot(data = window, pupil = mpupil, windows = F, geom = 'boxplot')
p

}
