% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_data.R
\name{run_functional_anova}
\alias{run_functional_anova}
\title{Run a functional ANOVA on a dataframe previously fitted with b-splines.}
\usage{
run_functional_anova(data, pupil, alpha = 0.05)
}
\arguments{
\item{data}{a Pupil_function_data object from create_functional_anova_data}

\item{pupil}{column name indicating pupil data to test}

\item{alpha}{an alpha level to be set for the ANOVA}
}
\value{
A Pupil_anova_data dataframe
}
\description{
This allows running of a functional ANOVA for a given alpha on pupil data that has been fitted with b-splines.
}
