% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smoothed_data.R
\name{replace_missing_data}
\alias{replace_missing_data}
\title{replaces missing observations if you have some degree of incomplete observations}
\usage{
replace_missing_data(data)
}
\arguments{
\item{data}{your data of class pupillometryR}
}
\value{
A time-stepped data frame
}
\description{
This is a useful function if you have a dataset where certain timepoints have been removed for whatever reason,
but you want continuous time data. This will make assumptions about trials being the same length though,
so may not be appropriate for all data types.
This should only be run after running make_pupillometry_data.
}
\examples{
data(pupil_data)
Sdata <- make_pupillometryr_data(data = pupil_data,
subject = ID,
trial = Trial,
time = Time,
condition = Type)
new_data <- replace_missing_data(data = Sdata)
}
